Change protocol default for profilecli
